refactor(): move shape uploading function to utils
